# cuda4dnn(conv): fuse eltwise with convolutions