## Changes:

- use `device::size_type` and `device::index_type` in `Span`
- refactor functors
  - each functor has a `Params` member type (replacement for varidiac templates)
- fuses eltwise with convolution for three topologies (shown below)
- cuDNN fused convolution when actiation is ReLU and is applied after the eltwise operation

<cut/>

**activation(conv + eltwise)**

Examples: ResNet, ...

```
--------------               ---------------
| Bias Layer |               | Convolution |
--------------               ---------------
      |                             |
      |       ----------------      |
      --------| Eltwise Sum  |-------
              ----------------
                     |
              ----------------
              |  Activation  |                 
              ----------------
                     |
```

**eltwise + activation(conv)**
Examples: YOLOv3, YOLOv4, ....

```
--------------               ---------------
| Bias Layer |               | Convolution |
--------------               ---------------
      |                             |
      |                      ----------------
      |                      |  Activation  |                 
      |                      ----------------
      |                             |
      |       ----------------      |
      --------| Eltwise Sum  |-------
              ----------------
                     |
             
```

**conv + eltwise**
```
--------------               ---------------
| Bias Layer |               | Convolution |
--------------               ---------------
      |                             |
      |       ----------------      |
      --------| Eltwise Sum  |-------
              ----------------
                     |
           
```

---

### Benchmarks

**Device:** RTX 2080 Ti

Model        | Mode          | Time (without this PR)  | Time (with this PR)
-------------- | --------------- | ------------------------------ | --------------------------
YOLOv3    | FP32          | 17.44ms                        | 16.92ms
YOLOv3    | FP16          | 7.87ms                          | 7.71ms
YOLOv4    | FP32          | 18.01ms                        | 17.64ms
YOLOv4    | FP16          | 9.90ms                          | 9.68ms

608  x 608 input for YOLOv3 and YOLOv4.

Timings for YOLOv4 does not include NMS.

**Darknet Reference:**

Model        | Mode          | Time
-------------- | --------------- | -----------
YOLOv4    | FP32          | 24.8ms
YOLOv4    | FP16          | 20.6ms

Average of five runs of timings reported by `./darknet detector test cfg/coco.data cfg/yolov4.cfg yolov4.weights data/dog.jpg
`

---

**Pending:**
- need some test which checks all possible fusion combinations

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under OpenCV (BSD) License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [ ] There is reference to original bug report and related work
- [ ] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [ ] The feature is well documented and sample code can be built with the project CMake

```
force_builders=Custom
buildworker:Custom=linux-4
build_image:Custom=ubuntu-cuda:18.04
```