@YashasSamaga 2020-05-29
There is a small performance regression (~8% for ResNet 50) on some devices in FP16 mode. cuDNN's fused convolution is slower than unfused convolution. Since cuDNN is closed source, it's not possible to tell in what situations or devices fused convolution is slower.

**Possible options:**
1. Ignore the regression and let #16900 handle it
2. Disable cuDNN fused convolution for FP16 here and enable it in #16900