@alalek 2020-05-29
Add an option to control this fusing, enable by default. Something like:
```
utils::getConfigurationParameterBool("OPENCV_DNN_CUDA_CONVOLUTION_FUSE", true)
```