@YashasSamaga 2020-06-20
It turns out that fusing bias addition or eltwise with tensor core based convolutions is slower than the unfused tensor core convolution. Only half precision target uses tensor core based convolutions.

So now the fusion is only enabled in fp16 target for devices with CC lower than 7.0.